refactor(ale_claw): Tier 4 — split sanitize passes into canonical_sanitize.py
Moves the leaf sanitize passes (repair_orphaned_pairs + its 4 helpers,
ensure_valid_ordering, drop_thinking_blocks, sanitize_thinking_signatures, the
reasoning-signature helpers, downgrade_openai_reasoning) and the SYNTHETIC/_SKIP
constants to canonical_sanitize.py. These passes are leaf — they touch only the
canonical TYPES (now imported from canonical_types) and each other, never the
adapters or policy — confirmed by grep. The orchestrator sanitize_items and the
format adapters stay in canonical.py and import the passes back (which doubles as
the re-export), so every `from .canonical import repair_orphaned_pairs /
_SKIP_SYNTHESIS_STOP_REASONS / ...` (context, agent_loop, __init__, tests) is
unchanged and context.py's shared _SKIP constant stays the same object.
canonical.py 1288 -> 799 LOC. Pure move; re-exports identical; full suite green
(928 passed, 3 skipped).
